requirements.txt
Django>=3.1
django-filter>2.3
mysqlclient==2.0.1
django-rest-swagger
djangorestframework
drf-yasg